+++ sentry-spring/src/main/java/io/sentry/spring/tracing/SentrySpanAdvice.java
@@ -6,7 +6,6 @@ import io.sentry.ISpan;
 import io.sentry.SpanStatus;
 import io.sentry.util.Objects;
 import java.lang.reflect.Method;
-import java.util.concurrent.atomic.AtomicReference;
 import org.aopalliance.intercept.MethodInterceptor;
 import org.aopalliance.intercept.MethodInvocation;
 import org.jetbrains.annotations.NotNull;
@@ -29,7 +28,7 @@ public class SentrySpanAdvice implements MethodInterceptor {
 
   @Override
   public Object invoke(final @NotNull MethodInvocation invocation) throws Throwable {
-    final ISpan activeSpan = resolveActiveSpan();
+    final ISpan activeSpan = hub.getSpan();
 
     if (activeSpan == null) {
       // there is no active transaction, we do not start new span
@@ -43,8 +42,8 @@ public class SentrySpanAdvice implements MethodInterceptor {
       final SentrySpan sentrySpan =
           AnnotationUtils.findAnnotation(mostSpecificMethod, SentrySpan.class);
       span.setDescription(resolveSpanDescription(targetClass, mostSpecificMethod, sentrySpan));
-      if (sentrySpan != null && !StringUtils.isEmpty(sentrySpan.op())) {
-        span.setOperation(sentrySpan.op());
+      if (sentrySpan != null && !StringUtils.isEmpty(sentrySpan.operation())) {
+        span.setOperation(sentrySpan.operation());
       }
       try {
         final Object result = invocation.proceed();
@@ -65,20 +64,5 @@ public class SentrySpanAdvice implements MethodInterceptor {
     return sentrySpan == null || StringUtils.isEmpty(sentrySpan.value())
         ? targetClass.getSimpleName() + "." + method.getName()
         : sentrySpan.value();
-  }
-
-  private @Nullable ISpan resolveActiveSpan() {
-    final AtomicReference<ISpan> spanRef = new AtomicReference<>();
-
-    hub.configureScope(
-        scope -> {
-          final ISpan span = scope.getSpan();
-
-          if (span != null) {
-            spanRef.set(span);
-          }
-        });
-
-    return spanRef.get();
   }
 }