+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -4,6 +4,7 @@ import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import java.util.List;
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public final class HubAdapter implements IHub {
@@ -154,7 +155,7 @@ public final class HubAdapter implements IHub {
 
   @Override
   public SentryId captureTransaction(SentryTransaction transaction, Object hint) {
-    return Sentry.captureTransaction(transaction, hint);
+    return Sentry.getCurrentHub().captureTransaction(transaction, hint);
   }
 
   @Override
@@ -177,5 +178,20 @@ public final class HubAdapter implements IHub {
   @Override
   public @Nullable SentryTraceHeader traceHeaders() {
     return Sentry.traceHeaders();
+  }
+
+  @Override
+  public void setSpanContext(final @NotNull Throwable t, final @NotNull SpanContext sc) {
+    Sentry.getCurrentHub().setSpanContext(t, sc);
+  }
+
+  @Override
+  public @Nullable SpanContext getSpanContext(final @NotNull Throwable ex) {
+    return Sentry.getCurrentHub().getSpanContext(ex);
+  }
+
+  @Override
+  public @Nullable ISpan getSpan() {
+    return Sentry.getCurrentHub().getSpan();
   }
 }