+++ sentry/src/main/java/io/sentry/config/PropertiesProvider.java
@@ -1,5 +1,9 @@
 package io.sentry.config;
 
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -12,6 +16,27 @@ public interface PropertiesProvider {
    */
   @Nullable
   String getProperty(@NotNull String property);
+
+  /**
+   * Resolves a map for a property given by it's name.
+   *
+   * @param property - the property name
+   * @return the map or empty map if not found
+   */
+  @NotNull
+  Map<String, String> getMap(final @NotNull String property);
+
+  /**
+   * Resolves a list of values for a property given by it's name.
+   *
+   * @param property - the property name
+   * @return the list or empty list if not found
+   */
+  @NotNull
+  default List<String> getList(final @NotNull String property) {
+    final String value = getProperty(property);
+    return value != null ? Arrays.asList(value.split(",")) : Collections.emptyList();
+  }
 
   /**
    * Resolves property given by it's name.