+++ sentry/src/test/java/io/sentry/MainEventProcessorTest.kt
@@ -4,6 +4,7 @@ import com.nhaarman.mockitokotlin2.mock
 import io.sentry.hints.ApplyScopeData
 import io.sentry.hints.Cached
 import io.sentry.protocol.SdkVersion
+import java.lang.RuntimeException
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -24,10 +25,11 @@ class MainEventProcessorTest {
                 version = "1.2.3"
             }
         }
-        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment"): MainEventProcessor {
+        fun getSut(attachThreads: Boolean = true, attachStackTrace: Boolean = true, environment: String? = "environment", tags: Map<String, String> = emptyMap()): MainEventProcessor {
             sentryOptions.isAttachThreads = attachThreads
             sentryOptions.isAttachStacktrace = attachStackTrace
             sentryOptions.environment = environment
+            tags.forEach { sentryOptions.setTag(it.key, it.value) }
             return MainEventProcessor(sentryOptions)
         }
     }
@@ -132,7 +134,7 @@ class MainEventProcessorTest {
     }
 
     @Test
-    fun `when processing an event and attach threads is disabled, threads should not be set`() {
+    fun `when attach threads is disabled, threads should not be set`() {
         val sut = fixture.getSut(attachThreads = false, attachStackTrace = false)
 
         var event = SentryEvent()
@@ -142,7 +144,7 @@ class MainEventProcessorTest {
     }
 
     @Test
-    fun `when processing an event and attach threads is enabled, threads should be set`() {
+    fun `when attach threads is enabled, threads should be set`() {
         val sut = fixture.getSut()
 
         var event = SentryEvent()
@@ -152,13 +154,23 @@ class MainEventProcessorTest {
     }
 
     @Test
-    fun `when processing an event and attach threads is disabled, but attach stacktrace is enabled, current thread should be set`() {
+    fun `when attach threads is disabled, but attach stacktrace is enabled, current thread should be set`() {
         val sut = fixture.getSut(attachThreads = false, attachStackTrace = true)
 
         var event = SentryEvent()
         event = sut.process(event, null)
 
         assertEquals(1, event.threads.count())
+    }
+
+    @Test
+    fun `when attach threads is disabled, but attach stacktrace is enabled and has exceptions, threads should not be set`() {
+        val sut = fixture.getSut(attachThreads = false, attachStackTrace = true)
+
+        var event = SentryEvent(RuntimeException("error"))
+        event = sut.process(event, null)
+
+        assertNull(event.threads)
     }
 
     @Test
@@ -194,6 +206,25 @@ class MainEventProcessorTest {
         val event = SentryEvent()
         sut.process(event, null)
         assertEquals("custom", event.environment)
+    }
+
+    @Test
+    fun `sets tags from SentryOptions`() {
+        val sut = fixture.getSut(tags = mapOf("tag1" to "value1", "tag2" to "value2"))
+        val event = SentryEvent()
+        sut.process(event, null)
+        assertEquals("value1", event.tags["tag1"])
+        assertEquals("value2", event.tags["tag2"])
+    }
+
+    @Test
+    fun `when event has a tag set with the same name as SentryOptions tags, the tag value from the event is retained`() {
+        val sut = fixture.getSut(tags = mapOf("tag1" to "value1", "tag2" to "value2"))
+        val event = SentryEvent()
+        event.setTag("tag2", "event-tag-value")
+        sut.process(event, null)
+        assertEquals("value1", event.tags["tag1"])
+        assertEquals("event-tag-value", event.tags["tag2"])
     }
 
     private fun generateCrashedEvent(crashedThread: Thread = Thread.currentThread()) = SentryEvent().apply {